chore: try rollup
