+++ bin/rollup.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+ROLLUP_CONFIG_PATH="rollup/rollup.config.js"
+rollup -c ${LERNA_ROOT_PATH}/${ROLLUP_CONFIG_PATH}
