+++ package.json
@@ -30,7 +30,8 @@
     "pub:patch": "RELEASE_TYPE=\"patch\" ./bin/publish.sh",
     "pub:minor": "RELEASE_TYPE=\"minor\" ./bin/publish.sh",
     "pub:major": "RELEASE_TYPE=\"major\" ./bin/publish.sh",
-    "release": "standard-version --no-verify --releaseCommitMessageFormat \"chore: release\""
+    "release": "standard-version --no-verify --releaseCommitMessageFormat \"chore: release\"",
+    "new_build": "./bin/new_build.sh"
   },
   "browserslist": {
     "production": [
@@ -57,8 +58,13 @@
     "@alfalab/icons-glyph": "^1.6.0",
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
     "@babel/plugin-proposal-optional-chaining": "^7.9.0",
+    "@babel/preset-typescript": "^7.9.0",
     "@commitlint/config-conventional": "^8.3.4",
     "@commitlint/config-lerna-scopes": "^8.3.4",
+    "@rollup/plugin-babel": "^5.0.0",
+    "@rollup/plugin-commonjs": "^11.1.0",
+    "@rollup/plugin-node-resolve": "^7.1.3",
+    "@rollup/plugin-typescript": "^4.1.1",
     "@storybook/addon-actions": "^5.3.8",
     "@storybook/addon-docs": "^5.3.17",
     "@storybook/addon-knobs": "^5.3.8",
@@ -116,15 +122,18 @@
     "react-scripts": "^3.3.1",
     "replace-in-file": "^5.0.2",
     "rimraf": "^3.0.2",
-    "rollup": "^2.10.0",
+    "rollup": "^2.10.2",
+    "rollup-plugin-dts": "^1.4.3",
     "rollup-plugin-multi-input": "^1.1.1",
-    "rollup-plugin-typescript2": "^0.27.1",
+    "rollup-plugin-postcss": "^3.1.1",
+    "rollup-plugin-ts": "^1.2.24",
     "shelljs": "^0.8.3",
     "standard-version": "^8.0.0",
     "stylelint": "^12.0.0",
     "stylelint-config-standard": "^20.0.0",
     "ts-jest": "^25.4.0",
     "ts-loader": "^6.2.1",
+    "tslib": "^2.0.0",
     "typescript": "~3.8.3"
   },
   "peerDependencies": {
