+++ tsconfig.json
@@ -2,6 +2,7 @@
     "compilerOptions": {
         "skipLibCheck": true,
         "target": "ES5",
+        "module": "commonjs",
         "jsx": "react",
         "rootDir": ".",
         "outDir": "dist",
